+++ packages/gateway/src/gateway-language-model-settings.ts
@@ -27,7 +27,6 @@ export type GatewayModelId =
   | 'amazon/nova-pro'
   | 'anthropic/claude-3-haiku'
   | 'anthropic/claude-3.5-haiku'
-  | 'anthropic/claude-3.7-sonnet'
   | 'anthropic/claude-haiku-4.5'
   | 'anthropic/claude-opus-4'
   | 'anthropic/claude-opus-4.1'