+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,22 @@
 
 ### Patch Changes
 
+- Updated dependencies [38ca8dc]
+- Updated dependencies [6d76710]
+  - ai@7.0.0-canary.133
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
+- Updated dependencies [eaf849f]
+- Updated dependencies [8565dcb]
+  - ai@7.0.0-canary.132
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
 - Updated dependencies [b67525f]
 - Updated dependencies [ca446f8]
 - Updated dependencies [bcacd48]